+++ tests/testthat/test_grouped_ggpiestats.R
@@ -3,7 +3,7 @@ context("grouped_ggpiestats")
 testthat::test_that(
   desc = "grouped_ggpiestats works",
   code = {
-    skip_on_cran()
+    testthat::skip_on_cran()
 
     #--------------------- only main variable -------------------------------